SpriteObjects

+++ src/main/java/com/example/simplengine/Camera/Camera.java
@@ -26,7 +26,7 @@ public final class Camera {
         return instance;
     }
 
-    public void setPosition(float x, float y) {
+    public void setPosition(double x, double y) {
         this.position.setX(x);
         this.position.setY(y);
     }
@@ -47,47 +47,47 @@ public final class Camera {
         return zoom;
     }
 
-    public void move(float dx, float dy) {
+    public void move(double dx, double dy) {
         position.setX(position.getX() + dx);
         position.setY(position.getY() + dy);
     }
 
-    public Vector2 worldToScreen(float worldX, float worldY, int screenWidth, int screenHeight) {
-        float screenX = (float) ((worldX - position.getX()) * zoom + screenWidth / 2.0f);
-        float screenY = (float) ((worldY - position.getY()) * zoom + screenHeight / 2.0f);
+    public Vector2 worldToScreen(double worldX, double worldY, int screenWidth, int screenHeight) {
+        double screenX = (worldX - position.getX()) * zoom + screenWidth / 2.0f;
+        double screenY = (worldY - position.getY()) * zoom + screenHeight / 2.0f;
         return new Vector2(screenX, screenY);
     }
 
-    public Vector2 screenToWorld(float screenX, float screenY, int screenWidth, int screenHeight) {
-        float worldX = (float) ((screenX - screenWidth / 2.0f) / zoom + position.getX());
-        float worldY = (float) ((screenY - screenHeight / 2.0f) / zoom + position.getY());
+    public Vector2 screenToWorld(double screenX, double screenY, int screenWidth, int screenHeight) {
+        double worldX = (screenX - screenWidth / 2.0f) / zoom + position.getX();
+        double worldY = (screenY - screenHeight / 2.0f) / zoom + position.getY();
         return new Vector2(worldX, worldY);
     }
 
-    public void centerOn(float x, float y) {
+    public void centerOn(double x, double y) {
         setPosition(x, y);
     }
 
     public void centerOn(Vector2 target) {
-        setPosition((float) target.getX(), (float) target.getY());
+        setPosition((double) target.getX(), (double) target.getY());
     }
 
     public void follow(double targetX, double targetY, double deltaTime) {
 
-        float lerpFactor = Math.min(1.0f, followSpeed * (float) deltaTime * 60.0f);
+        double lerpFactor = Math.min(1.0f, followSpeed * (double) deltaTime * 60.0f);
 
-        float newX = (float) (position.getX() + (targetX - position.getX()) * lerpFactor);
-        float newY = (float) (position.getY() + (targetY - position.getY()) * lerpFactor);
+        double newX = (double) (position.getX() + (targetX - position.getX()) * lerpFactor);
+        double newY = (double) (position.getY() + (targetY - position.getY()) * lerpFactor);
         
         setPosition(newX, newY);
     }
 
     public void follow(Vector2 target, double deltaTime) {
-        follow((float) target.getX(), (float) target.getY(), deltaTime);
+        follow(target.getX(), target.getY(), deltaTime);
     }
 
-    public void setFollowSpeed(float speed) {
-        this.followSpeed = Math.max(0.01f, Math.min(1.0f, speed));
+    public void setFollowSpeed(double speed) {
+        this.followSpeed = (float) Math.max(0.01f, Math.min(1.0f, speed));
     }
 
     public float getFollowSpeed() {
